@@ -13,6 +13,7 @@
 //   limitations under the License.
 
 using Etherna.BeehiveManager.NetClient.DtoModels;
+using Etherna.BeehiveManager.NetClient.InputModels;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -76,11 +77,11 @@ namespace Etherna.BeehiveManager.NetClient
         public async Task<BeeNodeDto> FindNodeOwnerOfPostageBatchAsync(string batchId) =>
             CurrentApiVersion switch
             {
-                ApiVersions.v0_3 => new BeeNodeDto(await client.ApiV0_3PostageBatchesNodeAsync(batchId).ConfigureAwait(false)),
+                ApiVersions.v0_3 => new BeeNodeDto(await client.ApiV0_3LoadBalancerBatchAsync(batchId).ConfigureAwait(false)),
                 _ => throw new InvalidOperationException()
             };
 
-        public async Task<IEnumerable<BeeNodeDto>> FindBeeNodesPinningContentAsync(string hash, bool requireAliveNodes = false) =>
+        public async Task<IEnumerable<BeeNodeDto>> FindNodesPinningResourceAsync(string hash, bool requireAliveNodes = false) =>
             CurrentApiVersion switch
             {
                 ApiVersions.v0_3 => (await client.ApiV0_3PinningNodesAsync(hash, requireAliveNodes).ConfigureAwait(false)).Select(n => new BeeNodeDto(n)),
@@ -115,17 +116,17 @@ namespace Etherna.BeehiveManager.NetClient
                 _ => throw new InvalidOperationException()
             };
 
-        public async Task<IEnumerable<string>> GetPinsByNodeAsync(string nodeId) =>
-            CurrentApiVersion switch
-            {
-                ApiVersions.v0_3 => await client.ApiV0_3NodesPinsGetAsync(nodeId).ConfigureAwait(false),
-                _ => throw new InvalidOperationException()
-            };
-
         public async Task<PinnedResourceDto> GetPinDetailsAsync(string nodeId, string hash) =>
             CurrentApiVersion switch
             {
                 ApiVersions.v0_3 => new PinnedResourceDto(await client.ApiV0_3NodesPinsGetAsync(nodeId, hash).ConfigureAwait(false)),
+                _ => throw new InvalidOperationException()
+            };
+
+        public async Task<IEnumerable<string>> GetPinsByNodeAsync(string nodeId) =>
+            CurrentApiVersion switch
+            {
+                ApiVersions.v0_3 => await client.ApiV0_3NodesPinsGetAsync(nodeId).ConfigureAwait(false),
                 _ => throw new InvalidOperationException()
             };
 
@@ -150,27 +151,28 @@ namespace Etherna.BeehiveManager.NetClient
                 _ => throw new InvalidOperationException()
             };
 
-        public Task NotifyNodeOfUploadedPinnedContentAsync(string id, string hash) =>
+        public Task NotifyNodeOfUploadedPinnedResourceAsync(string id, string hash) =>
             CurrentApiVersion switch
             {
                 ApiVersions.v0_3 => client.ApiV0_3NodesPinsUploadedAsync(id, hash),
                 _ => throw new InvalidOperationException()
             };
 
-        public Task<string> PinContentInNodeAsync(string hash, string? nodeId = null) =>
+        public Task<string> PinResourceInNodeAsync(string hash, string? nodeId = null) =>
             CurrentApiVersion switch
             {
                 ApiVersions.v0_3 => client.ApiV0_3PinningAsync(hash, nodeId),
                 _ => throw new InvalidOperationException()
             };
 
-        public async Task<BeeNodeDto> RegisterNewNodeAsync(string connectionScheme, int debugApiPort, int gatewayApiPort, string hostname) =>
+        public async Task<BeeNodeDto> RegisterNewNodeAsync(string connectionScheme, int debugApiPort, int gatewayApiPort, string hostname, NodeConfigInput nodeConfig) =>
             CurrentApiVersion switch
             {
                 ApiVersions.v0_3 => new BeeNodeDto(await client.ApiV0_3NodesPostAsync(new Generated.BeeNodeInput
                 {
                     ConnectionScheme = connectionScheme,
                     DebugApiPort = debugApiPort,
+                    EnableBatchCreation = nodeConfig.EnableBatchCreation,
                     GatewayApiPort = gatewayApiPort,
                     Hostname = hostname
                 }).ConfigureAwait(false)),
@@ -184,9 +186,22 @@ namespace Etherna.BeehiveManager.NetClient
                 _ => throw new InvalidOperationException()
             };
 
+        public Task ReuploadResourceToNetwork(string nodeId, string hash) =>
+            CurrentApiVersion switch
+            {
+                ApiVersions.v0_3 => client.ApiV0_3NodesStewardshipPutAsync(nodeId, hash),
+                _ => throw new InvalidOperationException()
+            };
+
+        public async Task<BeeNodeDto> SelectHealthyNodeAsync() =>
+            CurrentApiVersion switch
+            {
+                ApiVersions.v0_3 => new BeeNodeDto(await client.ApiV0_3LoadBalancerAsync().ConfigureAwait(false)),
+                _ => throw new InvalidOperationException()
+            };
+
         public async Task<BeeNodeDto> SelectLoadBalancedNodeForDownloadAsync(string hash) =>
             CurrentApiVersion switch
-
             {
                 ApiVersions.v0_3 => new BeeNodeDto(await client.ApiV0_3LoadBalancerDownloadAsync(hash).ConfigureAwait(false)),
                 _ => throw new InvalidOperationException()
@@ -196,6 +211,25 @@ namespace Etherna.BeehiveManager.NetClient
             CurrentApiVersion switch
             {
                 ApiVersions.v0_3 => await client.ApiV0_3PostageBatchesTopupAsync(batchId, amount).ConfigureAwait(false),
+                _ => throw new InvalidOperationException()
+            };
+
+        public Task UpdateNodeConfigAsync(string nodeId, NodeConfigInput nodeConfig) =>
+            CurrentApiVersion switch
+            {
+                ApiVersions.v0_3 => client.ApiV0_3NodesConfigAsync(
+                    nodeId,
+                    new Generated.UpdateNodeConfigInput
+                    {
+                        EnableBatchCreation = nodeConfig.EnableBatchCreation
+                    }),
+                _ => throw new InvalidOperationException()
+            };
+
+        public Task<bool> VerifyResourceAvailabilityFromNode(string nodeId, string hash) =>
+            CurrentApiVersion switch
+            {
+                ApiVersions.v0_3 => client.ApiV0_3NodesStewardshipGetAsync(nodeId, hash),
                 _ => throw new InvalidOperationException()
             };
     }
