@@ -1,4 +1,18 @@
-﻿using System;
+﻿//   Copyright 2022-present Etherna Sagl
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
+
+using System;
 
 namespace Etherna.BeehiveManager.NetClient.DtoModels
 {
@@ -17,9 +31,9 @@ namespace Etherna.BeehiveManager.NetClient.DtoModels
             NodeId = resource.NodeId;
             Status = resource.Status switch
             {
-                Generated.PinnedResourceStatusDto.InProgress => PinnedResourceStatusDto.InProgress,
-                Generated.PinnedResourceStatusDto.NotPinned => PinnedResourceStatusDto.NotPinned,
-                Generated.PinnedResourceStatusDto.Pinned => PinnedResourceStatusDto.Pinned,
+                Generated.PinnedResourceDtoStatus.InProgress => PinnedResourceStatusDto.InProgress,
+                Generated.PinnedResourceDtoStatus.NotPinned => PinnedResourceStatusDto.NotPinned,
+                Generated.PinnedResourceDtoStatus.Pinned => PinnedResourceStatusDto.Pinned,
                 _ => throw new InvalidOperationException()
             };
         }
